Created Writer class
