+++ src/main/java/milk/products/iot/SourCream.java
@@ -1,14 +1,13 @@
-package milk.lab3.iot;
+package milk.products.iot;
 
-public class Milk extends MilkProduct {
+public class SourCream extends MilkProduct {
     private double percentageOfFat;
-    private boolean lactoseFree;
 
-    public Milk(String nameOfMilkProduct, String origin, double volumeOrMass, double priceInUahPerKilo, boolean homeMade, double percentageOfFat, boolean lactoseFree) {
+
+    public SourCream(String nameOfMilkProduct, String origin, double volumeOrMass, double priceInUahPerKilo, boolean homeMade, double percentageOfFat) {
         super(nameOfMilkProduct, origin, volumeOrMass, priceInUahPerKilo, homeMade);
 
         this.percentageOfFat = percentageOfFat;
-        this.lactoseFree = lactoseFree;
 
     }
     public String getHeaders(){
@@ -19,3 +18,4 @@ public class Milk extends MilkProduct {
         return this.getNameOfMilkProduct() + ", " + this.getPriceInUahPerUnit()+ ", "+ this.getHomeMade();
     }
 }
+
